# Optimize foreach for arrays where possible